Authentication system
Uses Discord OAuth2 as a login for this service instead of having whitelisted ids in a text file.
Complete overhaul in the website's security, potentially has some vulnerabilities but overall this is closer to what the system would look like in public production than what we had before.

While the database is automated, the whitelist itself is STILL manual. Giving people serverside access requires you to manually modify the users.db file and change the whitelist integer from 0 to 1.